refactor(memory): rename DefaultAllocator to HeapAllocator and expose public header
- Rename DefaultAllocator class to HeapAllocator to align with Phase 1 roadmap
- Create public header include/axiom/memory/heap_allocator.hpp
- Rename src/memory/default_allocator.cpp to src/memory/heap_allocator.cpp
- Update CMakeLists.txt to reference heap_allocator.cpp
- Convert class implementation from private to public API
- All 291 tests pass, including 28 allocator tests

This change improves API clarity and aligns with the roadmap specification
(Issue #015: HeapAllocator implementation). The global allocator API
(getDefaultAllocator/setDefaultAllocator) remains unchanged for backward
compatibility.

Closes #57
